[Feat] #39 UIViewController extension 파일로 빼서 구현
