+++ Relay/Relay/AppDelegate.swift
@@ -32,17 +32,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
 }
-
-extension UIViewController {
-    // 키보드 내리기
-    func hideKeyboardWhenTapped() {
-        let tap: UITapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(dissmissKeyboard))
-        tap.cancelsTouchesInView = false
-        view.addGestureRecognizer(tap)
-    }
-    
-    @objc
-    func dissmissKeyboard() {
-        view.endEditing(true)
-    }
-}
